Split some code from tuple_wrapper.cpp into tuple_policy.hpp and add more documentation
